prepare things for srv
