--- cmd/tgnotifier/version.go
@@ -1,3 +1,0 @@
-package main
-
-const version = "v0.4.4"
