+++ internal/tgnotifier/version.go
@@ -1,0 +1,4 @@
+package tgnotifier
+
+// Version is a version of the tgnotifier.
+const Version = "v0.4.4"
